Raise NoInitialConditionsError in kappa generator
This way the behaviour of kappa generator is more like BNG,
and it is possible to detect the no-initial-conditions error
cleanly.
